agent-engine/requirements.txt
# Core web framework and server
fastapi==0.103.2
uvicorn==0.23.2
gunicorn>=20.1.0,<21.0.0
python-dotenv==1.0.0
Jinja2>=3.0.0,<4.0.0

# Shared Alchemist libraries are installed separately in Dockerfile

# AI and LLM dependencies
pydantic
typing-inspect>=0.8.0,<1.0.0
typing-extensions>=4.5.0,<5.0.0
langchain
openai
langchain-openai
tiktoken>=0.5.2,<1.0.0

# Storage
redis>=5.0.0,<6.0.0

# File processing
python-multipart==0.0.7

# HTTP and API
aiohttp>=3.8.0,<4.0.0
requests>=2.28.0,<3.0.0
httpx==0.28.0

# Authentication
PyJWT>=2.0.0,<3.0.0

# Utilities
tqdm>=4.65.0,<5.0.0

# OpenAPI processing
PyYAML==6.0.1
openapi-spec-validator
prance

# CORS support is already included in FastAPI
